refactor: validateOrderLineItems 함수를 orderValidator에서 사용하도록 변경
